requirements.txt
requests==2.27.1
PyMySQL==1.0.2
lazysdk==0.0.24
fastmongo==0.0.8
lazymedia==0.0.1
easydownload==0.0.4
pymediainfo==6.0.1